pyproject.toml
[tool.poetry]
name = "instagram2"
version = "0.1.0"
description = ""
authors = ["ImAgiD <daurbekoaajgerim@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
python-decouple = "^3.8"
django-filter = "^24.2"
drf-yasg = "^1.21.7"
django-jazzmin = "^3.0.0"
redis = "^5.0.4"
celery = "^5.4.0"
django-cors-headers = "^4.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
